Move agent utils into own package
