+++ agent/src/main/java/com/code_intelligence/jazzer/instrumentor/BUILD.bazel
@@ -11,6 +11,7 @@ kt_jvm_library(
     ],
     deps = [
         "//agent/src/main/java/com/code_intelligence/jazzer/runtime",
+        "//agent/src/main/java/com/code_intelligence/jazzer/utils",
         "@com_github_jetbrains_kotlin//:kotlin-reflect",
         "@maven//:org_ow2_asm_asm",
         "@maven//:org_ow2_asm_asm_commons",
