+++ agent/src/main/java/com/code_intelligence/jazzer/instrumentor/Hook.kt
@@ -19,6 +19,7 @@ package com.code_intelligence.jazzer.instrumentor
 import com.code_intelligence.jazzer.api.HookType
 import com.code_intelligence.jazzer.api.MethodHook
 import com.code_intelligence.jazzer.api.MethodHooks
+import com.code_intelligence.jazzer.utils.descriptor
 import java.lang.invoke.MethodHandle
 import java.lang.reflect.Method
 import java.lang.reflect.Modifier
