+++ agent/src/main/java/com/code_intelligence/jazzer/utils/BUILD.bazel
@@ -1,0 +1,10 @@
+load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "utils",
+    srcs = glob([
+        "*.java",
+        "*.kt",
+    ]),
+    visibility = ["//visibility:public"],
+)
